Read a glider file in netcdf IOOS format

This is a provisional function, written to handle provisional netcdf files created from the python pyglider processing chain as well as files downloaded from the Glider IOOS ERDDAP site.

read.glider.netcdf.ioos(file, debug)

Arguments

file

Name of a netcdf file.

debug

an integer specifying whether debugging information is to be printed during processing. If this is not provided, then the value of getOption("gliderDebug",0) is used. The printing is done by a call to gliderDebug. Setting debug=0 turns off this form of debugging, while higher values may yield more information, depending on the function. If one glider function calls another, it passes the value of debug but decreased by 1, which means that the value of debug controls not just the breadth of debugging, but also the depth.

Value

A glider object, i.e. one inheriting from glider. (This class inherits from oce::oce in the oce package.)

Details

The data are copied directly from the file, except that time is converted from an integer to a POSIX time. Variable names containing underscores are renamed as e.g. profile_direction to profileDirection, although the [[,glider-method mechanism works with either name, e.g. if g is a glider object, then g[["profileDirection"]] and g[["profile_direction"]] give the same result.

See also

Author

Chantelle Layton and Dan Kelley