Read a Slocum Glider file

These files do not use standard names for variables, but the nameMap argument facilitates renaming for storage in the returned object. (Renaming simplifies later analysis, e.g. permitting direct use of algorithms in the oce package, which assume that salinity is named "salinity", etc.) The original names of data items are retained in the metadata of the returned object, so that the [[ operator in the oce package can retrieve the data using either the original name (e.g. x[["sci_water_temp"]]) or the more standard name (e.g. x[["temperature"]]).

read.glider.slocum(
  file,
  debug,
  nameMap = list(conductivity = "sci_water_cond", temperature = "sci_water_temp",
    pressure = "sci_water_pressure", longitude = "lon", latitude = "lat", depth =
    "i_depth")
)

Arguments

file

A connection or a character string giving the name of the file to load.

debug

an integer specifying whether debugging information is to be printed during processing. If this is not provided, then the value of getOption("gliderDebug",0) is used. The printing is done by a call to gliderDebug. Setting debug=0 turns off this form of debugging, while higher values may yield more information, depending on the function. If one glider function calls another, it passes the value of debug but decreased by 1, which means that the value of debug controls not just the breadth of debugging, but also the depth.

nameMap

List used to rename data columns. See “Details”.

Value

An oce object holding the data, with variables renamed as described in “Details”, and with salinity added, as calculated by oce::swSCTp() which uses the UNESCO algorithm and assumes that the conductivity values are stored in S/m units.

See also

Author

Dan Kelley

Examples

library(oceanglider)
if (file.exists("~/slocum.csv")) {
    g <- read.glider.slocum("~/slocum.csv")
    summary(g)

    # 1. Plot time-depth trace, colour-coded for temperature
    par(mar=c(3, 3, 1, 1), mgp=c(2, 0.7, 0)) # thin margins
    cm <- colormap(z=g[['temperature']])
    drawPalette(colormap=cm, cex.axis=3/4)
    t <- g[["time"]]
    p <- g[["depth"]]
    plot(t, p, ylim=rev(range(p)), xlab="Time", ylab="Pressure [dbar]",
         col=cm$zcol, cex=1/2, pch=20)
    mtext(paste("Temperature, from", t[1]), cex=3/4)

    # 2. Plot distance-depth trace, colour-coded for temperature
    dist <- geodDist(g[['longitude']],g[['latitude']],alongPath=TRUE)
    par(mar=c(3, 3, 1, 1), mgp=c(2, 0.7, 0)) # thin margins
    cm <- colormap(z=g[['temperature']])
    drawPalette(colormap=cm, cex.axis=3/4)
    p <- g[["depth"]]
    plot(dist, p, ylim=rev(range(p)), xlab="Distance [km]", ylab="Pressure [dbar]",
         col=cm$zcol, cex=1/2, pch=20)
    mtext(paste("Temperature, from", t[1]), cex=3/4)
}