Read a glider file in a Slocum netcdf format

This is an unsupported and provisional function, written to handle a particular file created in Spring of 2023 by MUN researchers, as part of an OTN project.

read.glider.slocum.netcdf(
  file,
  nameMap = list(salinity = "glider_record/sci_rbrctd_salinity_00", SA =
    "absolute_salinity", temperature = "glider_record/sci_rbrctd_temperature_00", CT =
    "conservative_temperature", pressure = "glider_record/sci_rbrctd_seapressure_00",
    time = "glider_record/sci_rbrctd_timestamp", conductivity =
    "glider_record/sci_rbrctd_conductivity_00", latitude = "glider_record/m_gps_lat",
    longitude = "glider_record/m_gps_lon", profileLat = "profile_lat", profileLon =
    "profile_lon", density = "density", O2 = "oxygen_concentration", 
     u = "u", v =
    "v", vxi = "glider_record/m_initial_water_vx", vyi =
    "glider_record/m_initial_water_vy", vy = "glider_record/m_water_vy", vx =
    "glider_record/m_water_vx", id = "profile_id"),
  debug
)

Arguments

file

Name of a netcdf file.

nameMap

List used to rename data columns. The default is tailore to a particular file, and is subject to change without notice, as that file is explored.

debug

an integer specifying whether debugging information is to be printed during processing. If this is not provided, then the value of getOption("gliderDebug",0) is used. The printing is done by a call to gliderDebug. Setting debug=0 turns off this form of debugging, while higher values may yield more information, depending on the function. If one glider function calls another, it passes the value of debug but decreased by 1, which means that the value of debug controls not just the breadth of debugging, but also the depth.

Value

A glider object, i.e. one inheriting from glider. (This class inherits from oce::oce in the oce package.)

Details

The sample file does not use standard names for variables, but the nameMap argument facilitates renaming for storage in the returned object. (Renaming simplifies later analysis, e.g. permitting direct use of algorithms in the oce package, which assume that salinity is named "salinity", etc.) The original names of data items are retained in the metadata of the returned object, so that the [[ operator in the oce package can retrieve the data using either the original name (e.g. x[["sci_water_temp"]]) or the more standard name (e.g. x[["temperature"]]).

See also

Author

Dan Kelley with help from Chantelle Layton and Cameron Richardson.